requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
bleach==4.1.0
boto3==1.21.12
botocore==1.24.12
Brotli==1.0.9
cachetools==4.2.4
catboost==0.26.1
category-encoders==2.2.2
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.10.0
dask-kubernetes==2021.10.0
decorator==5.1.0
defusedxml==0.7.1
dice-ml==0.7.2
dill==0.3.4
distlib==0.3.4
distributed==2021.10.0
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-jsonfield==1.4.1
django-rest-passwordreset==1.2.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dowhy==0.6
econml==0.12.0
entrypoints==0.3
erroranalysis==0.1.28
fairlearn==0.7.0
filelock==3.4.2
Flask==1.0.4
Flask-Compress==1.10.1
Flask-Cors==3.0.9
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2021.11.1
fuzzywuzzy==0.18.0
gevent==20.9.0
google-auth==2.3.3
graphviz==0.19.1
greenlet==0.4.17
gunicorn==20.0.4
h5py==3.6.0
HeapDict==1.0.1
idna==3.3
imageio==2.13.5
imbalanced-learn==0.8.1
imblearn==0.0
install==1.3.5
interpret==0.2.5
interpret-community==0.22.0
interpret-core==0.2.5
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonfield==3.1.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kubernetes==21.7.0
kubernetes-asyncio==19.15.0
lightgbm==3.2.1
lime==0.2.0.1
llvmlite==0.36.0
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.4.3
missingpy==0.2.0
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.3
multidict==5.2.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5
nltk==3.6.7
notebook==6.4.6
numba==0.53.1
numpy==1.21.5
oauthlib==3.1.1
orjson==3.6.4
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathos==0.2.8
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
plotly==5.3.1
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.8.0
# psycopg2==2.9.3
# psycopg2-binary==2.8.4
ptyprocess==0.7.0
pudb==2021.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.6
python-Levenshtein==0.12.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
rai-core-flask==0.2.4
raiwidgets==0.11.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
responsibleai==0.11.0
rich==11.2.0
rsa==4.8
s3transfer==0.5.2
SALib==1.4.5
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
shap==0.39.0
six==1.16.0
sklearn==0.0
skope-rules==1.0.1
slicer==0.0.7
sortedcontainers==2.4.0
sparse==0.13.0
sqlparse==0.4.2
statsmodels==0.13.1
sympy==1.9
tabulate==0.8.9
tblib==1.7.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
textblob==0.15.3
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==2.0.0
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
treeinterpreter==0.2.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
xgboost==1.3.3
yarl==1.7.2
zict==2.0.0
zope.event==4.5.0
zope.interface==5.4.0
